create useSaveTimeRecords custom hook
